Moved shader headers
